fork?

+++ uploadfile.php
@@ -49,15 +49,25 @@ if(isset($_FILES['file'])){
 		file_put_contents($jsonfile, $json_string);
 		echo "Successfully send file to: ";
 		echo $chat_to;
-		sleep(300);
-		$json_string = file_get_contents($jsonfile);
-		$data = json_decode($json_string, true);
-		if(isset($data[$hash_filename])){
-			unset($data[$hash_filename]);
-			unlink("upload/".$hash_filename);
-			$json_string = json_encode($data);
-			file_put_contents($jsonfile, $json_string);
+		/*
+		$pid = pcntl_fork();
+		if($pid==-1){
+			die('subprocess error');
+		}else if($pid){
+			exit();
 		}
+		else{
+			sleep(300);
+			$json_string = file_get_contents($jsonfile);
+			$data = json_decode($json_string, true);
+			if(isset($data[$hash_filename])){
+				unset($data[$hash_filename]);
+				unlink("upload/".$hash_filename);
+				$json_string = json_encode($data);
+				file_put_contents($jsonfile, $json_string);
+			}
+		}
+		*/
 	}else{
 		print_r($errors);
 	}
